pyproject.toml
[tool.poetry]
name = "pptx-planner"
version = "0.1.0"
description = ""
authors = ["L33T-Snack3r <daniel_lim_professional@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.3"
openai = "^1.91.0"
google-genai = "^1.21.1"
ipykernel = "^6.29.5"
dotenv = "^0.9.9"
pillow = "^11.2.1"
html2image = "^2.0.7"
selenium = "^4.33.0"
webdriver-manager = "^4.0.2"
playwright = "^1.53.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
